got dma reads working on pcIIa
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@74 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/tms9914/write.c
@@ -97,9 +97,9 @@ IBLCL void bdPIOwrt(ibio_op_t *wrtop)
 { 
 	faddr_t		buf;
 	unsigned	cnt;
-	uint8		s1, s2;		/* software copies of HW status regs... */
+	int		s1, s2;		/* software copies of HW status regs... */
 	int		cfgbits;
-	uint8		lsb;		/* unsigned residual LSB */
+	uint8_t		lsb;		/* unsigned residual LSB */
 	char		msb;		/* signed residual MSB */
 extern  int eosmodes;
         int bytes=0;
